api/requirements.txt
# FastAPI Backend Requirements

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Background Tasks
celery==5.3.6
redis==5.0.1

# WebSocket
python-socketio==5.11.0
python-engineio==4.9.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP & Utils
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-cors==1.0.0

# Rate Limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Existing BACOWR dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
langcodes>=3.3.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
pyyaml>=6.0
jsonschema>=4.17.0
